+++ api/child-sections.ts
@@ -1,6 +1,6 @@
 import type { VercelRequest, VercelResponse } from "@vercel/node";
-import { db } from "./db";
-import { childSection } from "./db/schema";
+import { db } from "../backend/db";
+import { childSection } from "../backend/db/schema";
 import { eq } from "drizzle-orm";
 
 function isDemoRequest(req: VercelRequest): boolean {
