--- api/ping.ts
@@ -1,5 +1,0 @@
-import type { VercelRequest, VercelResponse } from '@vercel/node'
-
-export default function handler(req: VercelRequest, res: VercelResponse) {
-  res.status(200).json({ message: 'pong' })
-}
